// nexa-finance-api/src/types/transaction.types.ts
import type { TransactionType } from "@prisma/client";
import type { CategorySummary } from "./category.types";

export interface TransactionFilter {
  userId: string;
  date?: {
    gte: Date; //Greater Than or Equal
    lte: Date; //Less Than or Equal
  };
  type?: TransactionType;
  categoryId?: string;
}

export interface TransactionSummary {
  totalExpenses: number;
  totalIncomes: number;
  balance: number;
  expensesByCategory: CategorySummary[];
}